Refactor Dockerfile to use conda env to simulate local dev env as closely as possible in prod; also remove nginx from image.
The Dockerfile now installs py3.6 instead of py2.7 as in the previous
version. (I developed the Django app with py3.6). This means that
supervisord will not work in the current image (as it only plays well
with py < 3.0). Therefore I've decided to separate nginx out from the
container serving the Django app with uwsgi.