[Bug fix] Handle string value of STAGE env.
STAGE should have a value of 0 or 1 (integer) but instead has a value of '0' or '1'. Until I figure out if there's a way to set the value properly in the compose file, I am handling this issue by casting the value of STAGE to int.

Without this fix, the production build of the graphene-django backend will run with the stage config when it should run  with the production config.